+++ CHANGELOG.md
@@ -14,6 +14,9 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2026-02-05
+  * Updated PackingSolver dependency
+  * Added irregular packing "free" rotations
 ### 2026-02-03
   * Updated PackingSolver dependency
   * Fixed part list sorting